app: api communication fully implemented

+++ api/src/workout/models/overload.py
@@ -6,7 +6,7 @@ from django.db import models
 class Overload(BaseModel):
     exercise = models.ForeignKey('workout.Exercise', on_delete=models.CASCADE, related_name='overloads',
                                  verbose_name=_('Exercise'))
-    value = models.IntegerField(verbose_name=_('Value'), blank=True, null=True)
+    value = models.FloatField(verbose_name=_('Value'), blank=True, null=True)
     training_active = models.ForeignKey('workout.TrainingActive', on_delete=models.CASCADE, related_name='overloads',
                                         verbose_name=_('Training active'))
     order = models.IntegerField(verbose_name=_('Order'), blank=True, null=True)
